MFD: TWL6030: Add Support for GPADC
MFD: TWL6030: Add Support for GPADC

This patch adds support for the General Purpose A/D Convertor in
TWL6030.
This driver is based on the twl4030-madc driver written by Mikko Ylinen
and Filipe Balbi, and currently has a 'misc' device interface. It needs
to be moved to drivers/hwmon as per comments from the community on the
twl4030-madc driver.

Change-Id: I84568c847502e562f7a1656b845f94b00e6aaecf
Signed-off-by: Nishant Kamat <nskamat@ti.com>
Signed-off-by: Balaji T K <balajitk@ti.com>
Signed-off-by: Oleksandr Kozaruk <oleksandr.kozaruk@ti.com>