omap: dispc: force L3_1 CD to NOSLEEP when dispc module is active.
With the DSS FIFO optimizations, ramdom lockups and reboots are seen. It has
been identified that L3_1 CD is idling and not responding to the traffic
initiated by DSS.  The Workaround suggested by Hardware team is to keep the
L3_1 CD in NO_SLEEP mode, when DSS is active.  Once DSS module is switched to
idle mode, put L3_1 CD to HW_AUTO.

This patch was reverted by 169c5f071800863a44b69f8a9f66d36343c65093, but
a test case was found that reproduces the original problem.

Change-Id: I7acfcaa407bc7a2423fcc7a82f3503a6fbcee331
Signed-off-by: Akash Choudhari <akashc@ti.com>
Signed-off-by: Avinash.H.M. <avinashhm@ti.com>
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>
CC: Todd Poynor <toddpoynor@google.com>
CC: Colin Cross <ccross@android.com>
CC: Nishanth Menon <nm@ti.com>
CC: Girish S G <girishsg@ti.com>
CC: Mahesh <mahesh@ti.com>
CC: Dandawate Saket <dsaket@ti.com>
CC: Lajos Molnar <molnar@ti.com>