gpio: omap: clear level bits when switching to edge detect in idle
The gpio controller in omap4 can only wake from idle on edge
triggered lines.  The level triggered interrupts are converted
to edge triggered in idle.  It appears that the edge triggered
enable bits are ignored if the level triggered bit is also set,
so clear the level trigger bit in idle and restore it later.

If a level triggered interrupt is pending when the level detect
register is cleared the edge detect setting will not detect it,
and the interrupt will be lost.  After clearing the level detect
bits, manually check the datain register against the saved
level detect values, and abort the transition if the interrupt
is pending.

Change-Id: I43fbee728cb6ebc407b8c4430a1cd37165354dc6
Signed-off-by: Colin Cross <ccross@android.com>