ASoC: omap-mcasp: Relax S/PDIF clocks/frame constraint
The first step in the current McASP clock dividers calculation is
very strict with respect to the McASP FCLK being an integer multiple
of 128 (number of clocks required for each S/PDIF frame).

The clock divider calculation algorithm doesn't search for exact
matches, but for dividers which can produce clock drift within an
allowed PPM range. For this reason, rejecting McASP FCLK rates not
divisible by 128, reduces the space of rates the algorithm can work
on. By simply including the clocks per frame value in divider
calculation we give more flexibility to the algorithm.

Change-Id: Ic007f9fa9e776c2e5475a36b5444e2ec39c8b0bb
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>