OMAP4+: HSI: Prevent HSI clock disable while ACWAKE is high
OMAP4+: HSI: Prevent HSI clock disable while ACWAKE is high

If HSI clocks are disabled while ACWAKE is high, it is possible that ACWAKE
goes low unexpectedly.
This is not seen on Link Layers which do not comply with MIPI, because in this
case ACWAKE and CAWAKE are high at the same time, so HSI clocks are not
disabled.
But for Link Layers which comply with MIPI, CAWAKE will go low independtly of
CAWAKE, and HSI driver might cut the clocks while ACWAKE is high, which causes
CAWAKE to have an unknwon value.

Change-Id: Ib37982b3702ed64a917626d1a7407d558df1b49a
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>